break IdentityProtocol tests file into Identity and IdentityProtocol
